+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+
+## [0.8.3]
+### Fixed
+
+- Now captures `id` and `clientId` from server response
+- Updated config to include `clientId` (also schema `v0.0.2`)
+
 ## [0.8.2]
 ### Fixed